code/pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
]
